Docs: e-paper pipeline, standalone build, UC8179_EInk example; 1.4.0
* README: supported displays (UC8179 800x480 incl. Seeed reTerminal E1001), the new pieces
  (PackMono, UC8179, EInkGhost, PackRLE images, tools), an e-paper quick start, building
  with and without Arduino.
* docs/API.md: builds (Arduino / ESP-IDF / ESPHome / host), PackRLE images, PackMono,
  UC8179 (bus contract, refresh kinds, measured timings, border, memory), EInkGhost (wear
  model, plan/done loop), tools.
* examples/UC8179_EInk: canvas -> PackMono -> EInkGhost -> UC8179 on an ESP32-S3.
* UC8179.h: header comments brought up to date with the driven border and one data
  polarity.

All examples compile for ESP32-S3 (arduino-cli, esp32 3.3.6); host test passes with
-Wall -Wextra.

+++ tests/host_standalone.cpp
@@ -76,7 +76,7 @@ static void test_mono_export() {
 static void test_prle_image(const char *path, int w, int h) {
   // Round trip in C first.
   std::vector<uint8_t> idx((size_t)w * h);
-  for (size_t i = 0; i < idx.size(); i++) idx[i] = (uint8_t)((i / 7 + (i % w > w / 2)) % 3);
+  for (size_t i = 0; i < idx.size(); i++) idx[i] = (uint8_t)((i / 7 + ((int)(i % w) > w / 2)) % 3);
   std::vector<uint8_t> enc(h * (2 + PRLE_STRIDE(w)));
   size_t n = prle_image_encode_idx8(idx.data(), w, h, enc.data());
   std::vector<uint8_t> flat(h * PRLE_STRIDE(w));
